requirements.txt
anyconfig==0.13.0
arabic-reshaper==2.1.3
beautifulsoup4==4.11.1
brotlipy==0.7.0
certifi==2021.10.8
click==8.1.2
cycler==0.11.0
diffimg==0.2.3
editdistance==0.6.0
einops==0.2.0
filelock==3.6.0
Flask==2.1.1
fonttools==4.32.0
future==0.18.2
gdown==4.4.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
imageio==2.17.0
imgaug==0.4.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.2
lmdb==1.3.0
magickwand==0.2
MarkupSafe==2.1.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-service==2.4.0
munch==2.5.0
networkx==2.8
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==7.0.0
prefetch-generator==1.0.1
protobuf==3.20.0
pyDeprecate==0.3.2
pyparsing==3.0.8
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
Shapely==1.8.1.post1
sklearn==0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
straug==0.1.2
tensorboardX==2.5
threadpoolctl==3.1.0
tifffile==2022.4.8
torch==1.7.1
torchmetrics==0.8.1
torchvision==0.8.2
tqdm==4.64.0
typing==3.7.4.3
Wand==0.6.7
Werkzeug==2.1.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
